requirements.txt
wheel
pandas
yake
wikipedia
transformers
torch
scikit-learn
nltk
spacy
numpy
plotly
python-dotenv
tqdm
flask
gunicorn
pinecone